requirements.txt
selenium>=4.15
webdriver-manager>=4.0
python-dotenv>=1.0
requests>=2.31
